Fix bmp280 2

+++ devices/bmp280_sensor.c
@@ -121,14 +121,14 @@ BMP2_INTF_RET_TYPE bmp280_sensor_read_i2c (uint8_t register_address, uint8_t *ar
 {
     std_error_t *error = (std_error_t*)user_data;
 
-    return config.read_i2c_callback(BMP2_I2C_ADDR_PRIM, register_address, array, array_size, config.i2c_timeout_ms, error);
+    return config.read_i2c_callback(BMP2_I2C_ADDR_PRIM, (uint16_t)register_address, array, (uint16_t)array_size, config.i2c_timeout_ms, error);
 }
 
 BMP2_INTF_RET_TYPE bmp280_sensor_write_i2c (uint8_t register_address, const uint8_t *array, uint32_t array_size, void *user_data)
 {
     std_error_t *error = (std_error_t*)user_data;
 
-    return config.write_i2c_callback(BMP2_I2C_ADDR_PRIM, register_address, (uint8_t*)array, array_size, config.i2c_timeout_ms, error);
+    return config.write_i2c_callback(BMP2_I2C_ADDR_PRIM, (uint16_t)register_address, (uint8_t*)array, (uint16_t)array_size, config.i2c_timeout_ms, error);
 }
 
 void bmp280_sensor_delay_us (uint32_t period_us, void *user_data)
